make '\n' not a separator of the digest by default
FFUZZYPP_BLOCKDIGEST2_TERMS included '\n' as a separator.
This behavior is now disabled by default.

This macro is preserved and can be redefined to `',', '\n'` if you want
to handle '\n' as a separator. You can also redefine this macro to empty
value if you know that ',' is not given.